+++ packages/logic-core/src/data/events/onSymbolDisplay.ts
@@ -8,11 +8,17 @@ export interface SymbolDisplayHandler {
    * Controls whether a symbol should be visible in the grid.
    *
    * @param grid The grid that is being displayed.
+   * @param solution The solution grid, if it is available.
    * @param symbol The symbol that is being displayed.
    * @param editing Whether the grid is being edited.
    * @returns True if the symbol should be displayed, false otherwise. The symbol will not be displayed if any handler returns false.
    */
-  onSymbolDisplay(grid: GridData, symbol: Symbol, editing: boolean): boolean;
+  onSymbolDisplay(
+    grid: GridData,
+    solution: GridData | null,
+    symbol: Symbol,
+    editing: boolean
+  ): boolean;
 }
 
 export function handlesSymbolDisplay<T extends Instruction>(